Use cligj for plugin registration for #357.
