+++ rasterio/rio/main.py
@@ -1,32 +1,38 @@
-# main: loader of all the command entry points.
+"""
+Main click group for CLI
+"""
 
-import sys
-import traceback
 
+import logging
 from pkg_resources import iter_entry_points
+import sys
 
-from rasterio.rio.cli import BrokenCommand, cli
+import click
+import cligj
+import cligj.plugins
+
+import rasterio
 
 
-# Find and load all entry points in the rasterio.rio_commands group.
-# This includes the standard commands included with Rasterio as well
-# as commands provided by other packages.
-#
-# At a mimimum, commands must use the rasterio.rio.cli.cli command
-# group decorator like so:
-#
-#   from rasterio.rio.cli import cli
-#
-#   @cli.command()
-#   def foo(...):
-#       ...
+def configure_logging(verbosity):
+    log_level = max(10, 30 - 10*verbosity)
+    logging.basicConfig(stream=sys.stderr, level=log_level)
 
-for entry_point in iter_entry_points('rasterio.rio_commands'):
-    try:
-        entry_point.load()
-    except Exception:
-        # Catch this so a busted plugin doesn't take down the CLI.
-        # Handled by registering a dummy command that does nothing
-        # other than explain the error.
-        cli.add_command(
-            BrokenCommand(entry_point.name))
+
+@cligj.plugins.group(plugins=(
+        ep for ep in list(iter_entry_points('rasterio.rio_commands')) +
+                     list(iter_entry_points('rasterio.rio_plugins'))))
+@cligj.verbose_opt
+@cligj.quiet_opt
+@click.version_option(version=rasterio.__version__, message='%(version)s')
+@click.pass_context
+def main_group(ctx, verbose, quiet):
+
+    """
+    Rasterio command line interface.
+    """
+
+    verbosity = verbose - quiet
+    configure_logging(verbosity)
+    ctx.obj = {}
+    ctx.obj['verbosity'] = verbosity
